refactor: ♻️ refactor frontmatter handling
move static methods to separate utils module
